+++ ors-api/src/main/java/org/heigit/ors/api/responses/routing/JSONBasedIndividualRouteResponse.java
@@ -13,12 +13,14 @@
  * if not, see <https://www.gnu.org/licenses/>.
  */
 
-package org.heigit.ors.api.responses.routing.json;
+package org.heigit.ors.api.responses.routing;
 
 import org.heigit.ors.api.requests.routing.RouteRequest;
 import org.heigit.ors.api.responses.common.boundingbox.BoundingBox;
 import org.heigit.ors.api.responses.common.boundingbox.BoundingBoxFactory;
-import org.heigit.ors.api.responses.routing.IndividualRouteResponse;
+import org.heigit.ors.api.responses.routing.json.JSONExtra;
+import org.heigit.ors.api.responses.routing.json.JSONLeg;
+import org.heigit.ors.api.responses.routing.json.JSONSegment;
 import org.heigit.ors.common.DistanceUnit;
 import org.heigit.ors.exceptions.StatusCodeException;
 import org.heigit.ors.routing.RouteExtraInfo;
@@ -32,10 +34,10 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class JSONBasedIndividualRouteResponse extends IndividualRouteResponse {
+public abstract class JSONBasedIndividualRouteResponse extends IndividualRouteResponse {
     protected BoundingBox bbox;
 
-    public JSONBasedIndividualRouteResponse(RouteResult result, RouteRequest request) throws StatusCodeException {
+    protected JSONBasedIndividualRouteResponse(RouteResult result, RouteRequest request) throws StatusCodeException {
         super(result, request);
         if (request.hasUseElevation() && request.getUseElevation())
             includeElevation = true;
@@ -44,7 +46,7 @@ public class JSONBasedIndividualRouteResponse extends IndividualRouteResponse {
     }
 
     protected List<JSONSegment> constructSegments(RouteResult routeResult, RouteRequest request) {
-        List segments = new ArrayList<>();
+        List<JSONSegment> segments = new ArrayList<>();
         for (RouteSegment routeSegment : routeResult.getSegments()) {
             segments.add(new JSONSegment(routeSegment, request, routeResult.getSummary().getDistance()));
         }