+++ ors-api/src/main/java/org/heigit/ors/api/responses/routing/geojson/GeoJSONRouteResponse.java
@@ -33,6 +33,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class GeoJSONRouteResponse extends RouteResponse {
+    @Schema(description = "A GeoJSON 'FeatureCollection' representing the routes returned from the request")
     @JsonProperty("type")
     public final String type = "FeatureCollection";
 
@@ -64,8 +65,8 @@ public class GeoJSONRouteResponse extends RouteResponse {
     }
 
     @JsonProperty("features")
-    public List getRoutes() {
-        return routeResults;
+    public GeoJSONIndividualRouteResponse[] getRoutes() {
+        return routeResults.toArray(new GeoJSONIndividualRouteResponse[0]);
     }
 
     @JsonProperty("metadata")